All Update

+++ PIJ-DailyExcercises/Day3 (18 Oct 2023)/Day3_3_NOTFINISHED_Calculator.java
@@ -15,9 +15,9 @@ public class Day3_3_NOTFINISHED_Calculator {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Enter 1st number: ");
-        int num1 = scanner.nextInt();
+        double num1 = scanner.nextDouble();
         System.out.print("Enter 2nd number: ");
-        int num2 = scanner.nextInt();
+        double num2 = scanner.nextDouble();
         int mathChoice = 0;
 
         while (true) {
